+++ pkg/ingester/series.go
@@ -97,7 +97,7 @@ func (s *memorySeries) add(v model.SamplePair) error {
 		createdChunks.Inc()
 	}
 
-	chunks, err := s.head().add(v)
+	newChunk, err := s.head().add(v)
 	if err != nil {
 		return err
 	}
@@ -105,18 +105,13 @@ func (s *memorySeries) add(v model.SamplePair) error {
 	// If we get a single chunk result, then just replace the head chunk with it
 	// (no need to update first/last time).  Otherwise, we'll need to update first
 	// and last time.
-	if len(chunks) == 1 {
-		s.head().C = chunks[0]
-	} else {
-		s.chunkDescs = s.chunkDescs[:len(s.chunkDescs)-1]
-		for _, c := range chunks {
-			first, last, err := firstAndLastTimes(c)
-			if err != nil {
-				return err
-			}
-			s.chunkDescs = append(s.chunkDescs, newDesc(c, first, last))
-			createdChunks.Inc()
+	if newChunk != nil {
+		first, last, err := firstAndLastTimes(newChunk)
+		if err != nil {
+			return err
 		}
+		s.chunkDescs = append(s.chunkDescs, newDesc(newChunk, first, last))
+		createdChunks.Inc()
 	}
 
 	s.lastTime = v.Timestamp
@@ -235,13 +230,13 @@ func newDesc(c encoding.Chunk, firstTime model.Time, lastTime model.Time) *desc 
 // Add adds a sample pair to the underlying chunk. For safe concurrent access,
 // The chunk must be pinned, and the caller must have locked the fingerprint of
 // the series.
-func (d *desc) add(s model.SamplePair) ([]encoding.Chunk, error) {
+func (d *desc) add(s model.SamplePair) (encoding.Chunk, error) {
 	cs, err := d.C.Add(s)
 	if err != nil {
 		return nil, err
 	}
 
-	if len(cs) == 1 {
+	if cs == nil {
 		d.LastTime = s.Timestamp // sample was added to this chunk
 		d.LastUpdate = model.Now()
 	}