+++ pkg/querier/batch/chunk_test.go
@@ -48,13 +48,12 @@ func mkChunk(t require.TestingT, from model.Time, points int, enc promchunk.Enco
 	require.NoError(t, err)
 	ts := from
 	for i := 0; i < points; i++ {
-		pcs, err := pc.Add(model.SamplePair{
+		npc, err := pc.Add(model.SamplePair{
 			Timestamp: ts,
 			Value:     model.SampleValue(float64(ts)),
 		})
 		require.NoError(t, err)
-		require.Len(t, pcs, 1)
-		pc = pcs[0]
+		require.Nil(t, npc)
 		ts = ts.Add(step)
 	}
 	return chunk.NewChunk(userID, fp, metric, pc, model.Time(0), ts)